requirements.txt
python-dotenv==1.0.0

Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

psycopg2-binary==2.9.9

coverage==5.5
